backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
